Refactor alert handler
